+++ assembly/assembly/test/index.ts
@@ -10,3 +10,4 @@ import "./internal/node-test";
 import "./internal/reflected-value";
 import "./internal/strict-equality";
 import "./internal/traversal";
+import "./internal/vitest";