+++ assembly/assembly/vitest/expect.ts
@@ -1,0 +1,5 @@
+export {
+  Expectation,
+  expect,
+  NegatedExpectation,
+} from "../jest/expect";