+++ assembly/assembly/vitest/types.ts
@@ -1,0 +1,18 @@
+export {
+  AssertionFacade,
+  SuiteContext,
+  TestContext,
+} from "../internal/context";
+import { SuiteContext, TestContext } from "../internal/context";
+
+export type ModuleContext = TestContext;
+export type TestFn = (context: TestContext) => void;
+export type SuiteFn = (context: SuiteContext) => void;
+export type HookFn = (context: TestContext) => void;
+export type ModuleHookFn = (context: ModuleContext) => void;
+
+export class TestOptions {}
+
+export class SuiteOptions {}
+
+export class HookOptions {}